(PE-19049) Can remove parameters from pe.conf
...using the update_pe_conf function.  Since a null isn't meaningful
for a hocon lookup parameter in pe.conf (or at least, I can't think of
why it would be, at the moment), a nil can be used to remove a parameter
that we want to clean up from the file. It's possible I am overlooking
something tricky about nil, undef in hiera/lookup...it might be
applicable to a nodes.conf file where we wanted to clear a parameter on
a specific node that had been set in pe.conf, but this function only
applies to pe.conf, so I think this is acceptable for now.